Moved log filtering into a discrete BepInEx plugin
